+++ graphique/models.py
@@ -14,7 +14,7 @@ import pyarrow as pa
 import pyarrow.compute as pc
 import strawberry
 from strawberry import Info
-from strawberry.field import StrawberryField
+from strawberry.types.field import StrawberryField
 from .core import Column as C
 from .inputs import links
 from .scalars import Long, scalar_map, type_map